update cau truc project
